+++ docs/changelog.md
@@ -4,12 +4,20 @@ All notable changes to this project will be documented in this file.
 
 ## [4.5.10]
 
+### Added
+
+*   New `mention-role` command
+
 ### Fixed
 
 *   If a member is muted and leaves and returns, he will always be muted
 *   User game is now displayed in the userinfo command
 *   Fix `invitations` command
 *   Rep cooldown is now user based and not member based
+*   `steal` command was renamed to `rob`
+*   `minimize` command was renamed to `shorturl`
+*   `ban` command now accepts ids
+*   Fixes some other minor bugs
 
 ## [4.5.0]